Move the after delay transformation code
